bas-data: IMemory.read() returns int instead of byte.
- compatible with I/O interfaces.
- Returns -1 for END.
- no need to cast to int.

+++ bas/layer-3/bas-data/src/main/java/net/bodz/bas/data/mem/RandomAccessFileMemory.java
@@ -68,14 +68,14 @@ public class RandomAccessFileMemory
     }
 
     @Override
-    public byte read(int addr)
+    public int read(int addr)
             throws MemoryAccessException {
         try {
             file.seek(offset + addr);
             int b = file.read();
             if (b == -1)
                 throw new MemoryAccessException("end of file");
-            return (byte) b;
+            return b;
         } catch (IOException e) {
             throw new MemoryAccessException(e);
         }
